Separate test helpers and tests of helpers
Previously we had tests for Rails helpers, as well as helper modules
used to DRY-up tests, scattered over several directories with
overlapping names. This collects all the Rails helper tests into
test/helpers and all the helper modules into test/support and ensures
they use a consistent naming scheme of _helpers.rb.
