+++ test/helpers/batch_invitations_helper_test.rb
