+++ test/helpers/root_helper_test.rb
