+++ test/helpers/users_helper_test.rb
