+++ test/integration/user_agent_test.rb
@@ -1,8 +1,8 @@
 require 'test_helper'
-require 'helpers/password_support'
+require 'support/password_helpers'
 
 class UserAgentIntegrationTest < ActionDispatch::IntegrationTest
-  include PasswordSupport
+  include PasswordHelpers
 
   setup do
     @user = create(:user, name: "Normal User")
