+++ test/policies/application_policy_test.rb
@@ -1,8 +1,8 @@
 require 'test_helper'
-require 'helpers/policy_helper'
+require 'support/policy_helpers'
 
 class ApplicationPolicyTest < ActiveSupport::TestCase
-  include PolicyHelper
+  include PolicyHelpers
 
   %i[index edit update manage_supported_permissions users_with_access].each do |permission_name|
     context permission_name do
