+++ test/support/analytics_helpers.rb
