+++ test/support/confirmation_token_helpers.rb
@@ -1,4 +1,4 @@
-module ConfirmationTokenHelper
+module ConfirmationTokenHelpers
   def token_sent_to(user)
     token = Devise.friendly_token
     user.confirmation_token = token
