+++ test/support/policy_helpers.rb
@@ -1,4 +1,4 @@
-module PolicyHelper
+module PolicyHelpers
   def permit?(user, record, action)
     self.class.to_s.gsub(/Test/, '').constantize.new(user, record).public_send("#{action}?")
   end
